chore: renaming who-we-are components
