+++ crowdin/messages.json
@@ -3150,7 +3150,6 @@
   "-176261994": "Learn more about swap",
   "-1690471280": "For financial, the swap charge is calculated based on this formula:",
   "-341285068": "One standard lot for Forex = 100,000 units",
-  "-1033881248": "See our open positions",
   "-293057652": "Deriv is one of the world’s largest online brokers. We offer CFDs and other derivatives on forex, stocks & indices, cryptocurrencies, commodities, and derived to millions of registered users across the globe.",
   "-77779780": "From inception, our goal was to break free of the high commissions and clunky products offered by traditional brokers. Also, we aim to deliver a first-class experience to digitally inclined traders, regardless of the size of their accounts.",
   "-1468738308": "In a journey spanning more than 22 years, we have grown to over 2.5 million customers worldwide. But our mission has remained the same.",
@@ -3179,10 +3178,7 @@
   "-275117592": "withdrawals last month",
   "-654476117": "114M+",
   "-608807773": "trades last month",
-  "-2022759357": "Make trading accessible to anyone, anywhere",
-  "-1855417261": "Our leadership",
-  "-418008452": "Our principles are the framework for our decisions",
-  "-3618456": "Learn more about our principles",
+  "-1033881248": "See our open positions",
   "-1277783292": "Chief Operating Officer",
   "-1369524546": "Chief Information Officer",
   "-216767904": "Chief Administrative Officer",
@@ -3190,6 +3186,10 @@
   "-290006480": "Vice President of Product",
   "-871474425": "Vice President of IT Operations",
   "-864130274": "Head of Accounts",
+  "-2022759357": "Make trading accessible to anyone, anywhere",
+  "-1855417261": "Our leadership",
+  "-418008452": "Our principles are the framework for our decisions",
+  "-3618456": "Learn more about our principles",
   "-234962411": "Who we are | An Online Trading Platform | Deriv.com",
   "-719133900": "Deriv is a pioneering and award-winning online trading platform that offers a wide selection of derivatives for anyone, anywhere to trade.",
   "-328903026": "Best Forex Broker Africa",
