+++ src/pages/who-we-are/carousel/_carousel.tsx
