+++ src/pages/who-we-are/carousel/_image-marquee.tsx
