+++ src/pages/who-we-are/components/_make-trading.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import styled from 'styled-components'
-import { first_section_texts } from './_data'
+import { first_section_texts } from '../_data'
 import { localize } from 'components/localization'
 import { SectionContainer } from 'components/containers'
 import { Header } from 'components/elements'
