+++ src/pages/who-we-are/components/_our-offices.tsx
@@ -1,8 +1,8 @@
 import React from 'react'
 import styled from 'styled-components'
 import { graphql, useStaticQuery } from 'gatsby'
-import { desktop_pins, mobile_pins, our_offices_count } from './_data'
-import MapPin from './components/_map-pin'
+import { desktop_pins, mobile_pins, our_offices_count } from '../_data'
+import MapPin from './_map-pin'
 import { localize } from 'components/localization'
 import { SectionContainer, CssGrid, Desktop, Mobile, Flex } from 'components/containers'
 import { Header, Text, BackgroundImage } from 'components/elements'
