+++ src/pages/who-we-are/components/_our-principles.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import styled from 'styled-components'
-import { our_principles } from './_data'
+import { our_principles } from '../_data'
 import { localize } from 'components/localization'
 import { SectionContainer, Flex, Box } from 'components/containers'
 import { Header, Divider } from 'components/elements'
